Change/fix the way the mechanism for remotely starting hashpipe on a DAQ node.
The head node passes to start_daq.py:
    - the list of module IDs to expect data from
    - the max file size
    - the run dir name

start_daq.py then creates
    - module.config (list of module IDs)
    - a shell script to run hashpipe, passing it run dir name and max file size
It runs the shell script, uses "pgrep" to find the PID of the hashpipe process,
and writes that to a file.

This is the only way that Ryan and I could figure out to run hashpipe
where the python script exits, but hashpipe keeps running,
and another process can kill it later.

TODO:
- change start.py accordingly
- change hashpipe to not create run dir, and not create HK stuff