Problem: we required that all DAQ nodes have the same username and data dir. At Lick (currently the only place we have > 1 computer) the 2 computers have different usernames. So I decided go back to the earlier daq_config.json format, where each DAQ node has its own username and data dir.
The file also specifies the IP addr and data dir of the head node.
The lets us tell whether the head node is also a DAQ node.
The scripts complain and quit if run on a non-head node.

So we now support the following types of configurations:

1) single node: both head node and DAQ node.
I considered the idea of restructuring the DAQ operations
(*_daq.py) so that they could be done as function calls
(rather than invocation via SSH).
But this adds complexity and doesn't buy anything.

2) multiple nodes; head node is also a DAQ node.

3) multiple nodes: head node is not a DAQ node

I added some alternative daq_config files for Lick
to exercise these cases.